+++ src/main/java/edu/montana/gsoc/msusel/quamoco/processor/extents/ProjectNodeExtentDecorator.java
@@ -32,7 +32,6 @@ import edu.montana.gsoc.msusel.codetree.node.structural.ProjectNode;
 import edu.montana.gsoc.msusel.codetree.node.type.TypeNode;
 import edu.montana.gsoc.msusel.quamoco.model.NormalizationRange;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -51,19 +50,19 @@ public class ProjectNodeExtentDecorator extends AbstractNodeExtentDecorator {
     }
 
     @Override
-    public BigDecimal findFileExtent(String metric) {
+    public double findFileExtent(String metric) {
         ProjectNode p = (ProjectNode) decorated;
         return sumMetrics(metric, (List<FileNode>) p.files());
     }
 
     @Override
-    public BigDecimal findMethodExtent(String metric) {
+    public double findMethodExtent(String metric) {
         ProjectNode p = (ProjectNode) decorated;
         return sumMetrics(metric, (List<MethodNode>) p.methods());
     }
 
     @Override
-    public BigDecimal findClassExtent(String metric) {
+    public double findClassExtent(String metric) {
         ProjectNode p = (ProjectNode) decorated;
         return sumMetrics(metric, (List<TypeNode>) p.types());
     }
